+++ server/controllers/taskController.ts
@@ -4,8 +4,39 @@ import Project from "../models/projectModel";
 import ErrorResponse from "../utils/errorResponse";
 import { isValidObjectId } from "mongoose";
 
+// @desc Get tasks for a project
+// @route GET /api/task/
+// @access private
+export const getTasksForProject = async (req: Request, res: Response, next: NextFunction) => {
+  const { projectID } = req.body;
+  const { user } = <any>req;
+
+  //Checks if provided project id can be casted ot ObjectId
+  if (!isValidObjectId(projectID)) {
+    return next(new ErrorResponse("Invalid project ID", 400));
+  }
+
+  //Check if project exist
+  const project = await Project.findById(projectID);
+  if (!project) {
+    return next(new ErrorResponse("Project not found", 404));
+  }
+
+  //Check if user is the owner of the project
+  if (project.createdBy.toString() !== user._id.toString()) {
+    return next(new ErrorResponse("There was an error", 403));
+  }
+
+  try {
+    const tasks = await Task.find({ projectID });
+    res.status(200).json(tasks);
+  } catch (err) {
+    next(err);
+  }
+};
+
 // @desc Create new task
-// @route POST /task/
+// @route POST /api/task/
 // @access private
 export const createTask = async (req: Request, res: Response, next: NextFunction) => {
   const { title, desc, projectID } = req.body;
@@ -37,11 +68,12 @@ export const createTask = async (req: Request, res: Response, next: NextFunction
 };
 
 /// @desc Update task
-/// @route PUT /task/:id
+/// @route PUT /api/task/:id
 /// @access private
 export const updateTask = async (req: Request, res: Response, next: NextFunction) => {
   const { title, desc } = req.body;
   const { id } = req.params;
+  const { user } = <any>req;
 
   console.log(id);
 
@@ -51,8 +83,8 @@ export const updateTask = async (req: Request, res: Response, next: NextFunction
   }
 
   try {
-    const task = await Task.findByIdAndUpdate(
-      id,
+    const task = await Task.findOneAndUpdate(
+      { _id: id, createdBy: user._id },
       {
         title: title,
         desc: desc,
@@ -64,6 +96,11 @@ export const updateTask = async (req: Request, res: Response, next: NextFunction
       return next(new ErrorResponse("Project not found", 404));
     }
 
+    //Check if user is the owner of the task
+    if (task.createdBy.toString() !== user._id.toString()) {
+      return next(new ErrorResponse("There was an error", 401));
+    }
+
     res.status(200).json(task);
   } catch (err) {
     next(err);
@@ -71,10 +108,11 @@ export const updateTask = async (req: Request, res: Response, next: NextFunction
 };
 
 // @desc Delete task
-// @route DELETE /task/:id
+// @route DELETE /api/task/:id
 // @access private
 export const deleteTask = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params;
+  const { user } = <any>req;
 
   //Checks if provided task id can be casted ot ObjectId
   if (!isValidObjectId(id)) {
@@ -82,9 +120,7 @@ export const deleteTask = async (req: Request, res: Response, next: NextFunction
   }
 
   try {
-    const task = await Task.findOneAndDelete({
-      _id: id,
-    });
+    const task = await Task.findOneAndDelete({ _id: id, createdBy: user._id });
 
     //Checks if provided task id exists in database
     if (!task) {
